Moved several find functions from Mock to Transmon object
